Gnome-Keyring: Fix error message on system start
The pam_gnome_keyring.so module chooses a directory under
$XDG_RUNTIME_DIR with a random string in it and exports it
under the user session as $GNOME_KEYRING_CONTROL. The
"--start" flag allows to reuse this variable.

I'd like it more if both the GNOME_KEYRING_CONTROL directory
(and the DISPLAY) were persistent, but that's in the
(hopefully near!) future for now.